requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.2.1
Deprecated==1.2.18
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.1
Flask-Limiter==3.13
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rich==14.1.0
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
